frontend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==5.5.0
    # via
    #   frontend (pyproject.toml)
    #   streamlit
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.0
    # via
    #   google-auth
    #   streamlit
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via streamlit
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
google-api-core==2.25.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
google-auth==2.41.1
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-core
google-cloud-bigquery==3.38.0
    # via frontend (pyproject.toml)
google-cloud-core==2.4.3
    # via google-cloud-bigquery
google-crc32c==1.7.1
    # via google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-bigquery
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.75.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.75.1
    # via google-api-core
idna==3.10
    # via requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
narwhals==2.6.0
    # via altair
numpy==2.3.3
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==25.0
    # via
    #   altair
    #   google-cloud-bigquery
    #   streamlit
pandas==2.3.3
    # via
    #   frontend (pyproject.toml)
    #   streamlit
pillow==11.3.0
    # via streamlit
proto-plus==1.26.1
    # via google-api-core
protobuf==6.32.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   streamlit
pyarrow==21.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via
    #   google-cloud-bigquery
    #   pandas
pytz==2025.2
    # via pandas
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   streamlit
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.50.0
    # via frontend (pyproject.toml)
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.2
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   grpcio
    #   streamlit
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
watchdog==6.0.0
    # via streamlit
